+++ slatedb/src/compactor_state.rs
@@ -760,11 +760,12 @@ impl CompactorState {
                     v.insert(compaction.clone());
                 }
                 Entry::Occupied(mut o) => {
-                    let adopt = matches!(compaction.status(), CompactionStatus::Compacted)
-                        || (matches!(compaction.status(), CompactionStatus::Running)
-                            && matches!(o.get().status(), CompactionStatus::Scheduled))
-                        || (matches!(compaction.status(), CompactionStatus::Scheduled)
-                            && matches!(o.get().status(), CompactionStatus::Running));
+                    let adopt = (matches!(o.get().status(), CompactionStatus::Running)
+                        && matches!(compaction.status(), CompactionStatus::Compacted))
+                        || (matches!(o.get().status(), CompactionStatus::Scheduled)
+                            && matches!(compaction.status(), CompactionStatus::Running))
+                        || (matches!(o.get().status(), CompactionStatus::Running)
+                            && matches!(compaction.status(), CompactionStatus::Scheduled));
                     if adopt {
                         o.insert(compaction.clone());
                     } else {
